requirements.txt
vpmobil==2.1.0.1
pydantic-settings==2.13.1
pyyaml==6.0.3
flask==3.1.3
lxml==6.0.2
requests